+++ frontend/src/pages/Login.tsx
@@ -28,18 +28,20 @@ const Login = () => {
             username,
             password,
           }),
-        },
+        }
       );
 
       if (!response.ok) {
         const errorText = await response.text();
         throw new Error(
-          `Server error ${response.status}: ${errorText || "No details"}`,
+          `Server error ${response.status}: ${errorText || "No details"}`
         );
       }
 
       const data = await response.json();
       console.log(data.message);
+      localStorage.setItem("userId", data.userId);
+      localStorage.setItem("username", username);
 
       //localStorage.setItem('token', data.token);
       navigate("/");
@@ -57,7 +59,9 @@ const Login = () => {
         <h2 className="text-2xl font-bold text-white mb-4">Sign In</h2>
 
         <p
-          className={`text-sm text-red-600 mb-4 ${errorMessage ? "block" : "hidden"}`}
+          className={`text-sm text-red-600 mb-4 ${
+            errorMessage ? "block" : "hidden"
+          }`}
         >
           {errorMessage}
         </p>
